+++ src/api/axios.js
@@ -2,7 +2,8 @@ import axios from "axios";
 import { fetchAuthSession } from "@aws-amplify/auth";
 
 const api = axios.create({
-  baseURL: "https://r6zsml2v4c.execute-api.ap-southeast-1.amazonaws.com/PRD",
+  //baseURL: "https://r6zsml2v4c.execute-api.ap-southeast-1.amazonaws.com/PRD",
+   baseURL: "http://localhost:8688",
   headers: {
     "Content-Type": "application/json"
   }
